verbs: Add flow steering egress support
Sometimes we want to carry actions when packets are sent out to the
port. In that case, we need an EGRESS based flow steering rule.
The semantics of this new EGRESS flag is port-wide, meaning - any
packet that is sent to the port is matched, regardless of the QP that
this packet originated from. The flow steering QP parameter is used
only as a presentor of the device and if ignored in the matching
process.

issue: 1087238
Change-Id: Ifa9cabaa87676054ea8ea242a6fdfcd80fa46eff
Signed-off-by: Matan Barak <matanb@mellanox.com>